Currency

+++ Assets/Scripts/Scriptable/Minable.cs
@@ -6,6 +6,6 @@ public class Minable : ScriptableObject
 {
     public string _name;
     public float _health;
-    public int _minGather, _maxGather;
+    public long _minGather, _maxGather;
     public GameObject _prefab;
 }
